Major updates to function
The function was first updated so that the two cases where we have and don't have a stratification variable are dealt with the same piece of code. Done by introducing a dummy variable when no stratification variables specified.

The secondly the function now considers two cases when we want to run (a) on all areas and (b) aggregated from the reference area level. The former is needed to calculate the PYRS and the latter is needed for the likelihood of the function.